moved img and menu folders to respective components, added subfolders for components folder, added css modules to smaller components
