+++ frontend/package.json
@@ -6,7 +6,9 @@
   "dependencies": {
     "@testing-library/jest-dom": "^5.3.0",
     "@testing-library/react": "^10.0.2",
+    "@types/classnames": "^2.2.10",
     "babel-eslint": "^10.1.0",
+    "classnames": "^2.2.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-google-login": "^5.1.10",
