+++ frontend/src/App.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import LandingPage from './pages/Landing/Landing';
 import Dashboard from './pages/Dashboard/Dashboard';
-import { AuthManager } from './components/AuthManager';
+import { AuthManager } from './components/AuthManager/AuthManager';
 
-import './App.css';
+import './styles/App.css';
 
 const App = () => (
   <Router>
