+++ frontend/src/components/AuthManager/AuthManager.tsx
@@ -1,9 +1,9 @@
 import React, { useState, FunctionComponent } from 'react';
 import { GoogleLogin } from 'react-google-login';
 import { useHistory, useLocation } from 'react-router-dom';
-import useClientId from '../hooks/useClientId';
-import AuthContext from '../context/auth';
-import LandingPage from '../pages/Landing/Landing';
+import useClientId from '../../hooks/useClientId';
+import AuthContext from '../../context/auth';
+import LandingPage from '../../pages/Landing/Landing';
 
 export const AuthManager: FunctionComponent = ({ children }) => {
   const [signedIn, setSignedIn] = useState(false);
