+++ frontend/src/components/Logout/Logout.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, FunctionComponent } from 'react';
 import { GoogleLogout } from 'react-google-login';
-import useClientId from '../hooks/useClientId';
-import AuthContext from '../context/auth';
+import useClientId from '../../hooks/useClientId';
+import AuthContext from '../../context/auth';
 
 const Logout: FunctionComponent = () => {
   const clientId = useClientId();
