+++ frontend/src/components/Sidebar/icons/home.png
