+++ frontend/src/components/Sidebar/icons/index.ts
@@ -1,0 +1,4 @@
+export { default as drivers } from './drivers.png';
+export { default as home } from './home.png';
+export { default as riders } from './riders.png';
+export { default as settings } from './settings.png';
