+++ frontend/src/components/Sidebar/icons/riders.png
