+++ frontend/src/components/Sidebar/icons/settings.png
