+++ frontend/src/components/Sidebar/sidebar.module.css
@@ -1,4 +1,4 @@
-.sidebar-links{
+.sidebarLinks{
   text-decoration: none;
   color: black;
 }
