+++ frontend/src/pages/Dashboard/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
-import Sidebar from '../../components/Sidebar';
+import Sidebar from '../../components/Sidebar/Sidebar';
 import DriverTable from './DriversTable';
 import RiderTable from './RidersTable';
 import Home from './Home';
